refactored test data
